requirements.txt
# UOTA Elite v2 - 24/7 Cloud Operation Dependencies
# Core Framework
crewai>=0.75.0
crewai-tools>=0.17.0
langchain>=0.3.7
langchain-community>=0.3.5
langgraph>=0.2.45

# Trading & Exchange Integration
pandas>=2.2.3
numpy>=1.26.48
pandas-ta>=0.3.14b0
# Technical Analysis & ML
scikit-learn>=1.5.2
freqtrade>=2024.10

# Data & Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
aiofiles>=24.1.0
asyncio
schedule>=1.2.2

# Telegram Integration (Phone C2)
python-telegram-bot>=21.6
aiogram>=3.13.1

# API & Networking - High Performance
aiohttp>=3.11.2
requests>=2.32.3
websockets>=13.1
python-socketio>=5.11.4

# Utilities & Logging
python-dotenv>=1.0.1
loguru>=0.7.2
rich>=13.9.4
typer>=0.12.5
pydantic>=2.10.2
pydantic-settings>=2.6.1

# Time & Date
python-dateutil>=2.9.0.post0
pytz>=2024.2

# Configuration & Environment
pyyaml>=6.0.2
toml>=0.10.2

# Performance & Optimization - Zero Latency
numba>=0.60.0
cython>=3.0.11
uvloop>=0.21.0

# Security & Encryption - Military Grade
cryptography>=44.0.0
bcrypt>=4.2.1
pywin32>=308

# Monitoring & Metrics - 24/7 Operation
prometheus-client>=0.20.0
psutil>=6.1.0
winshell>=0.6

# Mathematical & Statistical
scipy==1.14.1
statsmodels==0.14.4
arch==7.0.0

# Image Processing (for chart analysis)
opencv-python==4.10.0.84
pillow==10.4.0
matplotlib>=3.9.2

# Async Support
asyncpg==0.30.0
redis==5.2.0

# Rate Limiting
ratelimit==2.2.1
slowapi==0.1.9

# Windows Specific - VPS Operation
pywin32>=308
winshell>=0.6
wmi>=1.5.1

# Process Management
psutil>=6.1.0
procfs>=0.9.0

# Network Optimization
aiofiles>=24.1.0
orjson>=3.10.11
msgpack>=1.1.0

# System Monitoring
GPUtil>=1.4.0
py-cpuinfo>=9.0.0
py-smbios>=0.4.0
